pyproject.toml
[tool.poetry]
name = "search-documents"
version = "0.1.0"
description = ""
authors = ["cehl-kurage <yusakuishihira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.0"
torch = {url = "https://download.pytorch.org/whl/cu117/torch-1.13.1%2Bcu117-cp310-cp310-linux_x86_64.whl"}
langchain = "^0.0.123"
openai = "^0.27.2"
faiss = {path = "/home/ocha/projects/document-search/build/faiss/python"}
python-dotenv = "^1.0.0"
hydra-core = "^1.3.2"
pyrootutils = "^1.0.4"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
